@@ -1,7 +1,5 @@
 package com.example.books_api.dtos;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,13 +10,11 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class BookDto {
-    // Create DTO (adding records)
-    @NotBlank(message = "title is required")
+public class UpdateBookDto {
+
     @Size(min = 2, max = 100, message = "Title must be between 2 and 100 characters")
     private String title;
 
-    @NotBlank(message = "author is required")
     @Size(min = 2, max = 100, message = "Author must be between 2 and 100 characters")
     private String author;
 
@@ -27,9 +23,8 @@ public class BookDto {
     private Integer publishYear;
 
     @Positive(message = "Page number must be positive")
-    private Integer pageNumber; // number of pages in a book
+    private Integer pageNumber;
 
-    @NotNull(message = "price is mandatory")
     @Positive(message = "Price must be greater than 0")
     private Double price;
 
