@@ -1,16 +1,19 @@
 package com.example.books_api.mapper;
 
 import com.example.books_api.dtos.BookDto;
+import com.example.books_api.dtos.BookResponseDto;
+import com.example.books_api.dtos.UpdateBookDto;
 import com.example.books_api.entities.Book;
-import org.mapstruct.BeanMapping;
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.NullValuePropertyMappingStrategy;
+import org.mapstruct.*;
 
 @Mapper(componentModel = "spring")
 public interface BookMapper {
     BookDto toDto(Book book);
     // This updates the existing Book entity with data from BookDto
+
+
+    @Mapping(source = "bookFile.fileName", target = "fileName")
+    BookResponseDto toResponseDto(Book entity);
     @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
-    void updateBookFromDto(BookDto dto, @MappingTarget Book entity);
+    void updateBookFromDto(UpdateBookDto dto, @MappingTarget Book entity);
 }
